se agrega modal para borrado y componentes

+++ src/components/CustomizedMenus.js
@@ -1,20 +1,20 @@
-import * as React from 'react';
-import IconButton from '@mui/material/IconButton';
-import Menu from '@mui/material/Menu';
-import MenuItem from '@mui/material/MenuItem';
-import MoreVertIcon from '@mui/icons-material/MoreVert';
+import * as React from "react";
+import IconButton from "@mui/material/IconButton";
+import Menu from "@mui/material/Menu";
+import MenuItem from "@mui/material/MenuItem";
+import MoreVertIcon from "@mui/icons-material/MoreVert";
 import { FaTrash, FaRegEdit } from "react-icons/fa";
-import { useDispatch } from 'react-redux';
-import { deleteProject, addProject } from "../features/projectSlice"; 
-
+import { useDispatch } from "react-redux";
+import { deleteProject, editProject } from "../features/projectSlice";
+import Swal from "sweetalert2";
 
 const ITEM_HEIGHT = 48;
 
 const iconStyles = {
-  marginRight: "12px"
-}
+  marginRight: "12px",
+};
 
-const CustomizedMenus = ({id}) => {
+const CustomizedMenus = ({ id }) => {
   const dispatch = useDispatch();
   const [anchorEl, setAnchorEl] = React.useState(null);
   const open = Boolean(anchorEl);
@@ -28,7 +28,21 @@ const CustomizedMenus = ({id}) => {
 
   const handleDelete = (id) => {
     setAnchorEl(null);
-    dispatch(deleteProject(id));
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteProject(id));
+        Swal.fire("Deleted!", "Your project has been deleted.", "success");
+      }
+    });
+
     console.log("borro ", id);
   };
 
@@ -41,8 +55,8 @@ const CustomizedMenus = ({id}) => {
       <IconButton
         aria-label="more"
         id="long-button"
-        aria-controls={open ? 'long-menu' : undefined}
-        aria-expanded={open ? 'true' : undefined}
+        aria-controls={open ? "long-menu" : undefined}
+        aria-expanded={open ? "true" : undefined}
         aria-haspopup="true"
         onClick={handleClick}
       >
@@ -51,7 +65,7 @@ const CustomizedMenus = ({id}) => {
       <Menu
         id="long-menu"
         MenuListProps={{
-          'aria-labelledby': 'long-button',
+          "aria-labelledby": "long-button",
         }}
         anchorEl={anchorEl}
         open={open}
@@ -59,21 +73,20 @@ const CustomizedMenus = ({id}) => {
         PaperProps={{
           style: {
             maxHeight: ITEM_HEIGHT * 4.5,
-            width: '20ch',
+            width: "20ch",
           },
         }}
       >
-        
-          <MenuItem  onClick={() => handleEdit(id)}>
-            <FaRegEdit style={iconStyles}/>Edit
-          </MenuItem>
-          <MenuItem  onClick={() => handleDelete(id)}>
-            <FaTrash style={iconStyles}/> Delete
-          </MenuItem>
-        
+        <MenuItem onClick={() => handleEdit(id)}>
+          <FaRegEdit style={iconStyles} />
+          Edit
+        </MenuItem>
+        <MenuItem onClick={() => handleDelete(id)}>
+          <FaTrash style={iconStyles} /> Delete
+        </MenuItem>
       </Menu>
     </div>
   );
-}
+};
 
-export default CustomizedMenus
+export default CustomizedMenus;
